remove ApplicationStatus and create approve()
