--- src/common/application-status.js
@@ -1,8 +1,0 @@
-export const ApplicationStatus = {
-  Approved: "APPROVED",
-  Received: "RECEIVED",
-  Offered: "OFFERED",
-  Accepted: "OFFER_ACCEPTED",
-  Rejected: "OFFER_REJECTED",
-  Withdrawn: "OFFER_WITHDRAWN",
-};
