+++ src/grants/events/application-approved.event.js
@@ -2,11 +2,12 @@ import { CloudEvent } from "../../common/cloud-event.js";
 
 export class ApplicationApprovedEvent extends CloudEvent {
   constructor(applicationApprovedEvent) {
+    const { code, clientRef, oldStatus, newStatus } = applicationApprovedEvent;
     super("application.approved", {
-      code: applicationApprovedEvent.code,
-      clientRef: applicationApprovedEvent.clientRef,
-      previousStatus: applicationApprovedEvent.previousStatus,
-      currentStatus: applicationApprovedEvent.currentStatus,
+      code,
+      clientRef,
+      oldStatus,
+      newStatus,
     });
   }
 }
