+++ src/grants/models/application.js
@@ -1,5 +1,4 @@
 import Boom from "@hapi/boom";
-import { ApplicationStatus } from "../../common/application-status.js";
 
 export const ApplicationPhase = {
   PreAward: "PRE_AWARD",
@@ -8,6 +7,15 @@ export const ApplicationPhase = {
 export const ApplicationStage = {
   Assessment: "ASSESSMENT",
   Award: "AWARD",
+};
+
+export const ApplicationStatus = {
+  Approved: "APPROVED",
+  Received: "RECEIVED",
+  Offered: "OFFERED",
+  Accepted: "OFFER_ACCEPTED",
+  Rejected: "OFFER_REJECTED",
+  Withdrawn: "OFFER_WITHDRAWN",
 };
 
 export class Application {
@@ -53,6 +61,11 @@ export class Application {
       answers,
       agreements: {},
     });
+  }
+
+  approve() {
+    this.status = ApplicationStatus.Approved;
+    this.updatedAt = this.#getTmestamp();
   }
 
   getFullyQualifiedStatus() {
