+++ src/grants/use-cases/accept-agreement.use-case.test.js
@@ -1,5 +1,4 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import { ApplicationStatus } from "../../common/application-status.js";
 import {
   Agreement,
   AgreementHistoryEntry,
@@ -9,6 +8,7 @@ import {
   Application,
   ApplicationPhase,
   ApplicationStage,
+  ApplicationStatus,
 } from "../models/application.js";
 import { CaseStatus } from "../models/case-status.js";
 import { publishApplicationStatusUpdated } from "../publishers/application-event.publisher.js";
