+++ src/grants/use-cases/submit-application.use-case.test.js
@@ -1,9 +1,9 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { ApplicationStatus } from "../../common/application-status.js";
 import {
   Application,
   ApplicationPhase,
   ApplicationStage,
+  ApplicationStatus,
 } from "../models/application.js";
 import { Grant } from "../models/grant.js";
 import { publishApplicationCreated } from "../publishers/application-event.publisher.js";
